% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledescription.R
\name{BaseDescription}
\alias{BaseDescription}
\title{\code{BaseDescription}}
\usage{
BaseDescription(description.of.n, n.total, n.subset, n.estimation, subset.label,
  weighted = TRUE, weight.label = "")
}
\arguments{
\item{description.of.n}{A description of the sample (e.g., "People").}

\item{n.total}{Total number of observations in the database.}

\item{n.subset}{Total number of observations in the subset (less than or
equal to \code{n.total}).}

\item{n.estimation}{The total number of observations used in estimation (less
than or equal to \code{subset}).}

\item{subset.label}{E.g., "Males living in New York".}

\item{weighted}{Whether sample has been weighted.}

\item{weight.label}{The label of the weight.}
}
\description{
Describes the sample, for use as footers in multivariate
  analyses.
}

