% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.R
\name{CloseBoldTagPlaceholder}
\alias{CloseBoldTagPlaceholder}
\title{Placeholder for html close bold tag </b>}
\usage{
CloseBoldTagPlaceholder()
}
\value{
Placeholder text for the close bold tag.
}
\description{
Insert this into table text in order to add the html close bold
tag, since </b> will be escaped if inserted directly.
}
