% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisontable.R
\name{ComparisonTable}
\alias{ComparisonTable}
\title{Creates a pretty formattable table for an comapring models.}
\usage{
ComparisonTable(values, order.values = FALSE, title = "",
  subtitle = "", footer = "")
}
\arguments{
\item{values}{A matrix of values.}

\item{order.values}{Specifies whether the values in the table are sorted by the first column.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}
}
\description{
Creates a pretty formattable table for an comapring models.
}
