% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ConvertCommaSeparatedStringToVector}
\alias{ConvertCommaSeparatedStringToVector}
\title{\code{ConvertCommaSeparatedStringToVector}}
\usage{
ConvertCommaSeparatedStringToVector(string)
}
\arguments{
\item{string}{A \code{\link{character}} to be converted.}
}
\value{
A \code{vector} of \code{character}s.
}
\description{
Converts a string containing commas into a vector, trimming whistepaces along the way.
}
