% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createanalysisreportwidget.R
\name{CreateAnalysisDesignWidget}
\alias{CreateAnalysisDesignWidget}
\title{Display widget for analysis reports for Q Trees}
\usage{
CreateAnalysisDesignWidget(x, css = NULL, ...)
}
\arguments{
\item{x}{A character vector with each element containing a string
in rtf format, an analysis report from tree model fit in Q or
Displayr.}

\item{css}{Optional string giving the path to a file containing
additional CSS to style the htmlwidget output.}

\item{...}{Currently, ignored.}
}
\value{
An \code{htmlwidget} containing the analysis report with
    some simple formatting.
}
\description{
Produces an \code{htmlwidget} containing the analysis report(s)
from a Mixed-Mode Trees, Latent Class Analysis, or Mixed-Mode
Cluster Analysis run in Q or Displayr.
}
\seealso{
\code{\link[striprtf]{read_rtf}},\code{\link[rhtmlMetro]{Box}}
}
