% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtextanalysiswidget.R
\name{CreateTextAnalysisWidget}
\alias{CreateTextAnalysisWidget}
\title{Display text analysis output as an \code{htmlwidget}}
\usage{
CreateTextAnalysisWidget(raw.and.normalized.text, n.gram.frequencies,
  token.substitutions, footer = "")
}
\arguments{
\item{raw.and.normalized.text}{A list containing the raw and normalized
text. The first element, called "Original Text", is a character vector of
the raw text. The second element, called "Transformed Text", is a list
containing the identified phrases for each row. The optional third
element, "Row Indices", contains the row indices of the text.}

\item{n.gram.frequencies}{A data frame with two variables, the first being
the n-gram and the second being the frequencies.}

\item{token.substitutions}{A character matrix with two columns mapping the
old tokens as they appeared in the original text (column 1) to the
normalized tokens (column 2).}

\item{footer}{Character; footer to show at the bottom of the output.}
}
\value{
An \code{htmlwidget} containing diagnostic information for
    the experimental design, including D-error, standard errors,
    frequenices, pairwise frequencies, the labeled design, and
    prior information. An attribute called \code{"ChartData"} also
    contains the labeled design.
}
\description{
Creates a \code{htmlwidget} summary of diagnostic information for a
text analysis output from flipTextAnalysis.
}
\seealso{
\code{\link[rhtmlMetro]{Box}}
}
