% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtextanalysiswidget.R
\name{CreateTransformedText}
\alias{CreateTransformedText}
\title{Extract the transformed text from the transformed tokens.}
\usage{
CreateTransformedText(x)
}
\arguments{
\item{x}{Either a object of class \code{wordBag} or a list with a
\code{transformed.text} element. In the former, the transformed text is created  from the transformed
tokenized list element. If the latter then the \code{transformed.text} element is returned.}
}
\value{
A character vector with the transformed text determined from the list of tokens
}
\description{
Convert the transformed tokenized element to transformed text. Adjusting for the blanks after
transformation as necessary.
}
