% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiontable.R
\name{CrosstabInteractionTable}
\alias{CrosstabInteractionTable}
\title{CrosstabInteractionTable}
\usage{
CrosstabInteractionTable(coef, coef.sign, group.size, footer = "",
  title = "", subtitle = "", sign.color = c("red", "black", "blue"),
  decimals = 2)
}
\arguments{
\item{coef}{Matrix containing estimates of the coefficient.}

\item{coef.sign}{Matrix of \code{-1,0,1}s indicating whether an interaction term makes a
coefficient significantly smaller (-1) or bigger (+1) than without the interaction.}

\item{group.size}{Vector containing the size of each group (i.e. interaction term).}

\item{footer}{Text to place in the footer of the table.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{sign.color}{Vector of three colors, which controls how coefficient values which are significant will be displayed.}

\item{decimals}{Number of decimals reported for the regression coefficients.}
}
\description{
Creates pretty table showing regression coefficients with a crosstab interaction.
}
