% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetmergingbycasewidget.R
\name{DataSetMergingByCaseWidget}
\alias{DataSetMergingByCaseWidget}
\title{Widget Output for Data Set Merging By Case}
\usage{
DataSetMergingByCaseWidget(
  input.data.sets.metadata,
  merged.data.set.metadata,
  matched.names,
  merged.names,
  omitted.variables,
  input.value.attributes,
  is.saved.to.cloud
)
}
\arguments{
\item{input.data.sets.metadata}{An object containing metadata for the
input data set.}

\item{merged.data.set.metadata}{An object containing metadata for the
merged data set.}

\item{matched.names}{A matrix whose rows contain the names of the input
variables used to create the merged variables. This matrix also has the
attributes "is.fuzzy.match" and "matched.by" which are matrices indicating
whether the input variable was a fuzzy match and what they were matched by,
i.e., variable name, label or value label.}

\item{merged.names}{A character vector of the names of the merged variables.
Also contains the attribute "renamed.variables" which contains the
merged variables that were renamed due to name conflicts and their
original names.}

\item{omitted.variables}{A list where each element contains the names of the
variables omitted from a data set.}

\item{input.value.attributes}{A list where each element contains the value
attributes of input categorical variables from a data set.}

\item{is.saved.to.cloud}{Whether the merged data set was saved to the
Displayr cloud drive.}
}
\description{
Widget shown in output for \code{flipData::MergeDataSetsByCase}.
}
