% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeplearningtable.R
\name{DeepLearningTable}
\alias{DeepLearningTable}
\title{DeepLearningTable}
\usage{
DeepLearningTable(
  values,
  column.labels = "Variable Importance",
  order.values = TRUE,
  title = "",
  subtitle = "",
  footer = ""
)
}
\arguments{
\item{values}{A named vector of values}

\item{column.labels}{Title of column}

\item{order.values}{Specifies whether the values in the table are sorted}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}
}
\description{
Creates a pretty formattable table for variable importance of a deep learning network.
}
\references{
This is based on code written by Kenton Russell.
}
