% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityextractionwidget.R
\name{EntityExtractionWidget}
\alias{EntityExtractionWidget}
\title{Display entity extraction output as an \code{htmlwidget}}
\usage{
EntityExtractionWidget(entity.percentages, variant.percentages,
  entity.counts, variant.counts, title, footer, empty.extraction)
}
\arguments{
\item{entity.percentages}{named numeric vector showing the percentage ocurrence of entity types
in the named entity recognition detection.}

\item{variant.percentages}{named numeric vector showing the percentage ocurrence of each
variant for each entity type in the named entity recognition detection.}

\item{entity.counts}{named numeric vector showing the raw counts of entity types in the named
entity recognition detection.}

\item{variant.counts}{named numeric vector showing the raw counts of each variant for each
entity type in the named entity recognition detection.}

\item{title}{The title to show at the top.}

\item{footer}{Footer to show containing sample information.}

\item{empty.extraction}{character giving the reason for a possible output with no entities. Returns
\code{NA} if entities extracted.}
}
\value{
An \code{htmlwidget} containing tables showing the output from an entity extraction.
}
\description{
Creates a \code{htmlwidget} summary of information for an
entity extraction output from flipTextAnalysis.
}
\seealso{
\code{\link[rhtmlMetro]{Box}}
}
