% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractchartdata.R
\name{ExtractChartData}
\alias{ExtractChartData}
\title{\code{ExtractChartData}}
\usage{
ExtractChartData(x)
}
\arguments{
\item{x}{Object containing the data. This is usually the output of a regression or machine learning output}
}
\description{
Returns data which is appropriate for charting
  The exact form of the data depends on the object
}
