% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExtractCommonPrefix}
\alias{ExtractCommonPrefix}
\title{\code{ExtractCommonPrefix}}
\usage{
ExtractCommonPrefix(labels)
}
\arguments{
\item{labels}{A vector of labels from which we plan to extract a common prefix.}
}
\value{
A \code{character}.
}
\description{
Returns the common prefix of a vector of labels.
}

