% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExtractCommonPrefixFromLabels}
\alias{ExtractCommonPrefixFromLabels}
\title{Extract shortened labels with common prefix removed}
\usage{
ExtractCommonPrefixFromLabels(data, tidy = TRUE)
}
\arguments{
\item{data}{data containing labels; see \code{\link{Labels}}}

\item{tidy}{logical; if \code{TRUE} the returned labels are also trimmed of whitespace
and have first letter capitalized}
}
\value{
A list with components
\itemize{
\item \code{common.prefix} - The common prefix shared among the labels,
or \code{NA} if none exists
\item \code{shortened.labels} - vector of labels with the common prefix removed.
}
}
\description{
Returns the common prefix of a vector of labels.
}
\seealso{
\code{\link{ExtractCommonPrefix}}, \code{\link{Labels}}, \code{\link{TidyLabels}}
}
