% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactbaseWidget.R
\name{FactbaseUploadWidget}
\alias{FactbaseUploadWidget}
\title{This function is for Displayr internal use only.}
\usage{
FactbaseUploadWidget(
  factbase.token = "",
  mode = "Replace all",
  aggregation = "None",
  time_aggregation = NULL,
  period_type = NULL,
  definition = NULL,
  hyperlink = NULL,
  owner = NULL,
  do.upload = TRUE,
  selection.type = "Table",
  input.table = NULL,
  make.dummy.metric = NULL,
  metric.variables = NULL,
  metric.column.names = "",
  date.column.name = "",
  date.variable = NULL,
  dimension.column.names = "",
  dimension.variables = NULL,
  output.type = "Data Summaries",
  start.date = "",
  time.zone = "Australia/Sydney",
  update.period = "days",
  update.frequency = 1,
  us.format = FALSE
)
}
\value{
An HTMLwidget that shows summary details about the upload.
}
\description{
The R code for the standard R item *Upload to Factbase*, which provides a
zero code way to upload a metric using the Factbase R API.
}
\details{
Parameters come out of the R UI JavaScript code in
https://github.com/Displayr/factbase/blob/master/displayr/Upload%20to%20Factbase.RScript
See that code to understand the meaning of those controls.

This is in this module because it relies on functions not exported by it:
createTempFile, createCata, addCss and boxIframeless.
}
