% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{FormatAsPercent}
\alias{FormatAsPercent}
\title{Converts a proportion to a percent}
\usage{
FormatAsPercent(x, digits = 2)
}
\arguments{
\item{x}{The number(s)}

\item{digits}{Number of significant digits.}
}
\description{
Converts a proportion to a percent
}
\details{
Multiplies by 100, keeping \code{digits} or more significant digits and
putting a % at the end, and commas if in thousands or more.
Based on \code{\link{formatC}}.
}

