% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{FormatAsPValue}
\alias{FormatAsPValue}
\title{\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05, and numbers greater than 0 are not rounded to 0,
unless is really close to 0.}
\usage{
FormatAsPValue(p, p.cutoff = 0.05, max.decimals = 12)
}
\arguments{
\item{p}{The number(s)}

\item{p.cutoff}{Ensures that values are not rounded to this value.}

\item{max.decimals}{When p is smaller than this value, it is returned as < this value (e.g., "< 0.001").}
}
\description{
\code{FormatAsPValue}
Formats p relatively nicely, ensuring that values greater than 0.05
are not rounded to 0.05, and numbers greater than 0 are not rounded to 0,
unless is really close to 0.
}
