% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{FormatAsReal}
\alias{FormatAsReal}
\title{Formats real numbers nicely.}
\usage{
FormatAsReal(x, digits = 2)
}
\arguments{
\item{x}{The number(s)}

\item{digits}{Number of significant digits}
}
\description{
Formats real numbers nicely.
}
\details{
Keeping \code{digits} or more significant digits and and commas if in thousands or more.
}
