% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{FormatWithDecimals}
\alias{FormatWithDecimals}
\title{Formats real numbers with a specified number of decimal places.}
\usage{
FormatWithDecimals(x, decimal.places = 2)
}
\arguments{
\item{x}{The number(s).}

\item{decimal.places}{Number of decimal places to show.}
}
\description{
Formats real numbers with a specified number of decimal places.
}

