% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{GetLabels}
\alias{GetLabels}
\title{GetLabels}
\usage{
GetLabels(names, data)
}
\arguments{
\item{names}{The names of the variables \code{data}.}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
A \code{vector} of names.
}
\description{
Replaces a list of given names, with any underlying labels, if they exist in the data.
}

