% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histtable.R
\name{HistTable}
\alias{HistTable}
\title{HistTable}
\usage{
HistTable(data.values, class.memberships = NULL, class.sizes = NULL,
  class.colors = NULL, title = "", subtitle = "", footer = "",
  bin.size = 5, bin.min = 0, bin.max = 100, hist.width = 100,
  hist.height = 20, show.tooltips = TRUE, color.negative = FALSE,
  histogram.column.name = "Distribution", ...)
}
\arguments{
\item{data.values}{A dataframe, with each column containing the values to create a histogram.}

\item{class.memberships}{Class memberships for each respondent.}

\item{class.sizes}{Sizes of each class as proportions.}

\item{class.colors}{Colors for each class.}

\item{title}{The title of the table.}

\item{subtitle}{The subtitle of the table.}

\item{footer}{The footer of the table.}

\item{bin.size}{Size of the bins used in the histogram. The breakpoints are given by \code{seq(bin.min, bin.max, bin.size)}.}

\item{bin.min}{Any value in \code{data.values} smaller then this will be truncated to \code{bin.min}.}

\item{bin.max}{Any value in \code{data.values} larger then this will be truncated to \code{bin.max}.}

\item{hist.width}{Width of the histogram cell in any valid CSS size unit}

\item{hist.height}{Height of the histogram cell}

\item{show.tooltips}{Whether to display tooltips of the bar heights}

\item{color.negative}{Whether to show negative bars in coral.}

\item{histogram.column.name}{Name for the histogram column.}

\item{...}{Additional columns to add to the table.}
}
\description{
Creates a pretty formattable table showing histograms
}
\examples{
dat <- data.frame(A=rpois(500,5), B=rpois(500,50), C=rpois(500,20))
print(HistTable(dat, 'Mean Probability'=c(5,50,100)))
}
