% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancetable.R
\name{ImportanceTable}
\alias{ImportanceTable}
\title{Create formattable table of importance scores}
\usage{
ImportanceTable(importance, row.labels, title = "", subtitle = "",
  footer = "", p.cutoff = 0.05)
}
\arguments{
\item{importance}{The importance object.}

\item{row.labels}{The row labels corresponding to the predictor variable names/labels.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}
}
\description{
Creates a pretty formattable table for relative/Shapley importance output.
}
\references{
This is based on code written by Kenton Russell.
}
