% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{Label<-}
\alias{Label<-}
\title{\code{"Label<-"}}
\usage{
Label(x) <- value
}
\arguments{
\item{x}{The object.}

\item{value}{A \code{character} object to set as a value.}
}
\description{
Assign a "label" attribute to an object.
}

