% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{Labels<-}
\alias{Labels<-}
\title{\code{"Label<-"}}
\usage{
Labels(x) <- value
}
\arguments{
\item{x}{An object or a list.}

\item{value}{A \code{character} object to set as a value or a vector of values.}
}
\description{
Assign a "label" attribute to an object.
}
