% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{Labels}
\alias{Labels}
\title{\code{Labels}}
\usage{
Labels(names, data)
}
\arguments{
\item{names}{The names of the variables or coefficients (e.g., Q2Cola for a factor).}

\item{data}{A \code{\link{data.frame}}.}
}
\value{
A \code{vector} of names.
}
\description{
Replaces a list of given names or coefficient, with any underlying labels, if they exist in the data.
Works for dummy variables as well as normal variables. Trims backticks and whitespace.
}

