% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdifftable.R
\name{MaxDiffTableClasses}
\alias{MaxDiffTableClasses}
\title{MaxDiffTableClasses}
\usage{
MaxDiffTableClasses(probabilities, column.labels, title = "",
  subtitle = "", footer = "")
}
\arguments{
\item{probabilities}{A matrix of the probabilities to be shown in the table.}

\item{column.labels}{Labels to show in the columns.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}
}
\description{
Creates a pretty formattable table for max-diff output.
}
\references{
This is based on code written by Kenton Russell.
}
