% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meancomparisonstable.R
\name{MeanComparisonsTable}
\alias{MeanComparisonsTable}
\title{MeanComparisonsTable}
\usage{
MeanComparisonsTable(means, zs, ps, r.squared, overall.p, column.names, footer,
  title = "", subtitle = "", p.cutoff = 0.05)
}
\arguments{
\item{means}{The means to be shown on the table.}

\item{zs}{Z-Statistics. Only used to determine the color of the fonts.}

\item{ps}{P-Values associated with each mean. Used to determine whether or not to color the cells.}

\item{r.squared}{The r-squared value for the regression predicting the row variable by the columns.}

\item{overall.p}{The P-value for the regression (e.g., an ANOVA F-Test for a linear model).}

\item{column.names}{The names to put on the columns of the table (other than for the R-Squared and overall P).}

\item{footer}{Text to place in the footer of the table.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}
}
\description{
Creates a pretty formattable table for mean comparisons.
}
\references{
This is based on code written by Kenton Russell.
}
