% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{Names}
\alias{Names}
\title{\code{Names}}
\usage{
Names(x)
}
\arguments{
\item{x}{A variable or a \link{data.frame}.}
}
\description{
Similar to \link{name}, except returns \code{attr(x, "name")} or \link{OriginalName} with a vector, if they exist.
}

