% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable.R
\name{OpenBoldTagPlaceholder}
\alias{OpenBoldTagPlaceholder}
\title{Placeholder for html open bold tag <b>}
\usage{
OpenBoldTagPlaceholder()
}
\value{
Placeholder text for the open bold tag.
}
\description{
Insert this into table text in order to add the html open bold
tag, since <b> will be escaped if inserted directly.
}
