% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{OriginalName}
\alias{OriginalName}
\title{\code{OriginalName}}
\usage{
OriginalName(x)
}
\arguments{
\item{x}{An object of some kind.}
}
\value{
A \code{character} of the name.
}
\description{
Finds the original name of an argument.
}
\references{
From http://stackoverflow.com/a/26558733/1547926.
}

