% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaloadingstable.R
\name{PCALoadingsTable}
\alias{PCALoadingsTable}
\title{PCALoadingsTable}
\usage{
PCALoadingsTable(
  loadings.matrix,
  variance.explained,
  eigenvalues,
  min.display.value,
  title = "",
  subtitle = "",
  footer = "",
  eigenvalue.label = "Eigenvalue"
)
}
\arguments{
\item{loadings.matrix}{A matrix of the PCA loadings/structure.}

\item{variance.explained}{A vector of the variance explained of the components.}

\item{eigenvalues}{A vector of the eigenvalues corresponding to the components.}

\item{min.display.value}{The minimum magnitude necessary for a value to be displayed.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{eigenvalue.label}{The label to display next to the eigenvalues.}
}
\description{
Creates a pretty formattable table for PCA loadings.
}
\references{
This is based on code written by Kenton Russell.
}
