% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appearancetemplate.R
\name{print.AppearanceTemplate}
\alias{print.AppearanceTemplate}
\title{Display appearance template as a \code{htmlwidget}}
\usage{
\method{print}{AppearanceTemplate}(x)
}
\arguments{
\item{x}{An object of class \code{AppearanceTemplate}.}
}
\description{
Display appearance template as a \code{htmlwidget}
}
