% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significancetest.R
\name{print.SignificanceTest}
\alias{print.SignificanceTest}
\title{print.SignificanceTest}
\usage{
\method{print}{SignificanceTest}(x, ...)
}
\arguments{
\item{x}{A \link{SignificanceTest} object}

\item{...}{Additional parameters to \code{\link{print.SignificanceTest}}}
}

