% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforesttable.R
\name{RandomForestTable}
\alias{RandomForestTable}
\title{RandomForestTable}
\usage{
RandomForestTable(importance, z.statistics, p.values,
  sort.by.importance = TRUE, title = "", subtitle = "", footer = "",
  p.cutoff = 0.05)
}
\arguments{
\item{importance}{A matrix of importance values from the regression.}

\item{z.statistics}{A vector or matrix of z statistics of the importance values.}

\item{p.values}{A vector or matrix of p values of the importance values.}

\item{sort.by.importance}{Sort the last column in descending order.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when determining significance.}
}
\description{
Creates a pretty formattable table for random forest.
}
\references{
This is based on code written by Kenton Russell.
}
