% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeimportancetable.R
\name{RelativeImportanceTable}
\alias{RelativeImportanceTable}
\title{RelativeImportanceTable}
\usage{
RelativeImportanceTable(relative.importance, raw.importance, standard.errors,
  t.statistics, p.values, title = "", subtitle = "", footer = "",
  p.cutoff = 0.05)
}
\arguments{
\item{relative.importance}{The relative importance scores.}

\item{raw.importance}{The raw importance scores.}

\item{standard.errors}{Standard errors of the raw importance scores.}

\item{t.statistics}{t-statistics for the raw importance scores.}

\item{p.values}{p-values for the raw importance scores.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}
}
\description{
Creates a pretty formattable table for relative importance output.
}
\references{
This is based on code written by Kenton Russell.
}

