% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{RemoveParentName}
\alias{RemoveParentName}
\title{\code{RemoveParentName}}
\usage{
RemoveParentName(x)
}
\arguments{
\item{x}{A \code{\link{character}}, typically containing a variable name.}
}
\value{
A \code{character}.
}
\description{
Replaces a list of given names, with any underlying labels, if they exist in the data.
}
\details{
Strips out everything up to and including the last \$. E.g., replaces
\code{"foo$fog$x"} with \code{"x"}
}
\examples{
RemoveParentName("foo$fog$x")
}

