% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReplacingEverythingBefore}
\alias{ReplacingEverythingBefore}
\title{\code{RemoveParentName}}
\usage{
ReplacingEverythingBefore(x, marker, include.marker = TRUE)
}
\arguments{
\item{x}{A \code{\link{character}} containing text to be modified.}

\item{marker}{A \code{character} containing the things to be searched for. You need to use
backslashes if not matching for letters (e.g., a \code{"\\$"}. Where there are mutliple instances, the last is found.}

\item{include.marker}{Whether or not the \code{marker} itself is to be replaced.}
}
\value{
A \code{character}.
}
\description{
Strips out everything up to and perhaps including a specific character.
}
\examples{
ReplacingEverythingBefore("foo$fog$x", "\\\\$")
ReplacingEverythingBefore("foo$fog$x", "\\\\$", FALSE)
ReplacingEverythingBefore("GiraffeDogCat", "Dog", FALSE)
ReplacingEverythingBefore("GiraffeDogCat", "Dog", TRUE)
}
