% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledescription.R
\name{SampleDescription}
\alias{SampleDescription}
\title{\code{SampleDescription}}
\usage{
SampleDescription(
  n.total,
  n.subset,
  n.estimation,
  subset.label,
  weighted = TRUE,
  weight.label = "",
  missing,
  imputation.label = NULL,
  m,
  variable.description = "",
  resample = FALSE,
  effective.sample.size = NULL
)
}
\arguments{
\item{n.total}{Total number of observations in the database.}

\item{n.subset}{Total number of observations in the subset (less than or
equal to \code{n.total}).}

\item{n.estimation}{The total number of observations used in estimation (less
than or equal to \code{subset}).}

\item{subset.label}{E.g., "Males living in New York".}

\item{weighted}{Whether sample has been weighted.}

\item{weight.label}{The label of the weight.}

\item{missing}{How missing data is to be treated in the analysis. Options
are: \code{"Error if missing data"}, \code{"Exclude cases with missing
data"}, ,and \code{"Imputation (replace missing values with estimates)"}.}

\item{imputation.label}{Method used to impute the data.}

\item{m}{Number of imputaton samples.}

\item{variable.description}{A \code{char} specifying the type of
variables being imputed. Used in the description of the imputation
that typically appears in a footer.}

\item{resample}{Whether resampling is used whenever weights are applied.}

\item{effective.sample.size}{The effective sample size when weights are applied.}
}
\value{
character description of sample
}
\description{
Describes the sample, for use as footers in multivariate
  analyses.
}
