% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significancetest.R
\name{SignificanceTest}
\alias{SignificanceTest}
\title{\code{SignificanceTest}}
\usage{
SignificanceTest(obj, var1, var2, filter, weight, show.labels = TRUE,
  decimal.places = NULL)
}
\arguments{
\item{obj}{Significance testing output object, e.g. object of class htest.}

\item{var1}{Variable 1 used in test.}

\item{var2}{Variable 2 used in test.}

\item{filter}{Filter variable.}

\item{weight}{Weight variable.}

\item{show.labels}{Whether to show variable labels instead of names.}

\item{decimal.places}{The number of decimal places to show.}
}
\description{
Standardize information to be shown in significance test results.
}

