% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{specifyDecimal}
\alias{specifyDecimal}
\title{specifyDecimal}
\usage{
specifyDecimal(x, decimals = 2, comma.for.thousands = TRUE)
}
\arguments{
\item{x}{The number(s).}

\item{decimals}{Number of decimal places to show.}

\item{comma.for.thousands}{If TRUE, uses a comma when there are thousands.}
}
\description{
Taken in part from https://stackoverflow.com/a/12135122/1547926https://stackoverflow.com/a/12135122/1547926
}
