% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingwidget.R
\name{StackingWidget}
\alias{StackingWidget}
\title{Stacking Output Widget}
\usage{
StackingWidget(
  input.data.set.metadata,
  stacked.data.set.metadata,
  unstackable.names,
  common.labels.list,
  is.saved.to.cloud
)
}
\arguments{
\item{input.data.set.metadata}{An object containing metadata for the
input data set.}

\item{stacked.data.set.metadata}{An object containing metadata for the
stacked data set.}

\item{unstackable.names}{A list of character vectors of the names of
variables that could not be stacked due to mismatching types or
value attributes.}

\item{common.labels.list}{A list of sets of common labels used for stacking.}

\item{is.saved.to.cloud}{Whether the stacked data set was saved to the
Displayr cloud drive.}
}
\description{
Stacking widget shown in output for \code{flipData::StackData}.
}
