% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TidyLabels}
\alias{TidyLabels}
\title{\code{TidyLabels}}
\usage{
TidyLabels(labels)
}
\arguments{
\item{labels}{A vector of labels from which we plan to extract a common prefix.}
}
\value{
A \code{character}.
}
\description{
Takes a vector of labels, removes any common prefix,  capitalizes the first letter, removes whitespace from the end.
}
