% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TrimLeadingWhitepsace}
\alias{TrimLeadingWhitepsace}
\title{\code{TrimLeadingWhitepsace}}
\usage{
TrimLeadingWhitepsace(x)
}
\arguments{
\item{x}{A \code{\link{character}} that may contain whitecontaining text to be modified.}
}
\value{
A \code{character}.
}
\description{
Removes whitespace (e.g.,spaces, tab characters) from the beginning of a string.
}

