% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turftable.R
\name{TURFTable}
\alias{TURFTable}
\title{TURFTable}
\usage{
TURFTable(output.table, portfolios, alternative.names, title = "",
  subtitle = "", footer = "")
}
\arguments{
\item{output.table}{A matrix of the statistics to be shown in the table.}

\item{portfolios}{A matrix containing the alternative indices of the TURF
portfolios.}

\item{alternative.names}{The alternative names.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}
}
\description{
Creates a pretty formattable table for TURF output.
}
