% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{VariableLabels}
\alias{VariableLabels}
\title{\code{VariableLabels}
Replaces a list of given names, with any underlying labels, if they exist in the data.}
\usage{
VariableLabels(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}
}
\value{
A \code{vector} of labels or names.
}
\description{
\code{VariableLabels}
Replaces a list of given names, with any underlying labels, if they exist in the data.
}

