% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceexplainedtable.R
\name{VarianceExplainedTable}
\alias{VarianceExplainedTable}
\title{VarianceExplainedTable}
\usage{
VarianceExplainedTable(
  eigenvalues,
  variance.proportions,
  cumulative.proportions,
  title = "",
  subtitle = "",
  footer = ""
)
}
\arguments{
\item{eigenvalues}{A vector of eigenvalues corresponding to the components.}

\item{variance.proportions}{A vector of the variance explained of the components.}

\item{cumulative.proportions}{A vector of the cumulative variance
explained of the components.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}
}
\description{
Creates a pretty formattable table for variance explained.
}
\references{
This is based on code written by Kenton Russell.
}
