% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecolours.R
\name{ZScoresAsColors}
\alias{ZScoresAsColors}
\title{\code{ZScoresAsColors}}
\usage{
ZScoresAsColors(
  z,
  light.negative.color = "#FB9080",
  dark.negative.color = "#FA614B",
  light.positive.color = "#80B4F4",
  dark.positive.color = "#3E7DCC",
  non.significant.color = "#888888",
  max.abs = 5,
  alpha = 0.05
)
}
\arguments{
\item{z}{A \code{\link{numeric}} vector containing colors.}

\item{light.negative.color}{The color, in standard sRGB colorspace, representing the least negative value.}

\item{dark.negative.color}{The color, in standard sRGB colorspace, representing the most negative value.}

\item{light.positive.color}{The color, in standard sRGB colorspace, representing the least positive value.}

\item{dark.positive.color}{The color, in standard sRGB colorspace, representing the most positive value.}

\item{non.significant.color}{The color used for non-significant values.}

\item{max.abs}{The (absolute) value beyond which all colors are set to the \code{negative.color} or \code{positive.color}}

\item{alpha}{The level of significance, where values not significant at this value are shaded with the \code{non.significant.color}.}
}
\description{
Shades z-scores.
}
