% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcustomtable.R
\name{CreateCustomTable}
\alias{CreateCustomTable}
\title{Display a table with custom formatting}
\usage{
CreateCustomTable(x, format.as.percentage = FALSE, format.decimals = 0,
  transpose = FALSE, global.font.family = "Arial",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  font.unit = "pt", values.fill = "transparent",
  values.border.width = 1, values.border.color = "#FFFFFF",
  values.align.horizontal = "center", values.align.vertical = "middle",
  values.font.family = global.font.family,
  values.font.color = global.font.color, values.font.size = 10,
  values.font.weight = "normal", values.font.style = "normal",
  colhead.show = TRUE, colhead.labels = NULL,
  colhead.fill = "#DDDDDD", colhead.border.width = 1,
  colhead.border.color = "#FFFFFF",
  colhead.align.horizontal = "center",
  colhead.align.vertical = "middle",
  colhead.font.family = global.font.family,
  colhead.font.color = global.font.color, colhead.font.size = 10,
  colhead.font.weight = "bold", colhead.font.style = "normal",
  rowhead.show = TRUE, rowhead.labels = NULL,
  rowhead.fill = colhead.fill, rowhead.border.width = 1,
  rowhead.border.color = "#FFFFFF", rowhead.align.horizontal = "left",
  rowhead.align.vertical = "middle",
  rowhead.font.family = global.font.family,
  rowhead.font.color = global.font.color, rowhead.font.size = 10,
  rowhead.font.style = "normal", rowhead.font.weight = "bold")
}
\arguments{
\item{x}{Input table to display.}

\item{format.as.percentage}{Display table values as percentage (multiply by 100
and add percentage sign). Ignored if \code{x} is not numeric.}

\item{format.decimals}{Controls number of decimal places shown in table cells.
Ignored if \code{x} is not numeric.}

\item{transpose}{Whether to switch rows and columns in \code{x}.}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute in the table unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{font.unit}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}

\item{values.fill}{Background color of the cells in the table.}

\item{values.border.width}{Width of border around table cells (in pixels).}

\item{values.border.color}{Color of border around table cells,}

\item{values.align.horizontal}{Horizontal alignment of text in table cells.}

\item{values.align.vertical}{Vertical alignment of text in table cells.}

\item{values.font.family}{Font family of text in table cells.}

\item{values.font.color}{Font color of text in table cells.}

\item{values.font.size}{Font size (in pixels) of text in table cells.}

\item{values.font.weight}{One of "normal" or "bold".}

\item{values.font.style}{One of "normal" or "italic".}

\item{colhead.show}{Logical; whether to show column headers in the table.
This will be ignored if \code{x} does not contain column names.}

\item{colhead.labels}{A vector or comma-separated labels to override the
column names of \code{x}.}

\item{colhead.fill}{Background color of the column headers in the table.}

\item{colhead.border.width}{Width of border around table column headers (in pixels).}

\item{colhead.border.color}{Color of border around table column headers,}

\item{colhead.align.horizontal}{Horizontal alignment of text in table column headers.}

\item{colhead.align.vertical}{Vertical alignment of text in table column headers.}

\item{colhead.font.family}{Font family of text in table column headers.}

\item{colhead.font.color}{Font color of text in table column headers.}

\item{colhead.font.size}{Font size (in pixels) of text in table column headers.}

\item{colhead.font.weight}{One of "normal" or "bold".}

\item{colhead.font.style}{One of "normal" or "italic".}

\item{rowhead.show}{Logical; whether to show row headers in the table.
This will be ignored if \code{x} does not contain row names.}

\item{rowhead.labels}{A vector or comma-separated labels to override
the column names of \code{x}.}

\item{rowhead.fill}{Background color of the row headers in the table.}

\item{rowhead.border.width}{Width of border around table row headers (in pixels).}

\item{rowhead.border.color}{Color of border around table row headers,}

\item{rowhead.align.horizontal}{Horizontal alignment of text in table row headers.}

\item{rowhead.align.vertical}{Vertical alignment of text in table row headers.}

\item{rowhead.font.family}{Font family of text in table row headers.}

\item{rowhead.font.color}{Font color of text in table row headers.}

\item{rowhead.font.size}{Font size (in pixels) of text in table row headers.}

\item{rowhead.font.style}{One of "normal" or "italic".}

\item{rowhead.font.weight}{One of "normal" or "bold".}
}
\description{
Displays a table which fills up the entire containiner with
 resizable column widths and row heights.
}
