% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtextanalysiswidget.R
\name{CreateTextAnalysisWidget}
\alias{CreateTextAnalysisWidget}
\title{Display text analysis output as an \code{htmlwidget}}
\usage{
CreateTextAnalysisWidget(raw.and.normalized.text, n.gram.frequencies,
  token.substitutions, footer, colors = NULL)
}
\arguments{
\item{raw.and.normalized.text}{A data frame with two variables, the first
being the raw text and the second being the normalized text. The row
names of the data frame should correspond to the case index.}

\item{n.gram.frequencies}{A data frame with two variables, the first being
the n-gram and the second being the frequencies.}

\item{token.substitutions}{A character matrix with two columns mapping the
old tokens as they appeared in the original text (column 1) to the
normalized tokens (column 2).}

\item{footer}{Character; footer to show at the bottom of the output.}

\item{colors}{Character; a vector containing colors for each entry in \code{n.gram.frequencies}.
The vector will be recycled if there is not enough values.}
}
\value{
An \code{htmlwidget} containing diagnostic information for
    the experimental design, including D-error, standard errors,
    frequenices, pairwise frequencies, the labeled design, and
    prior information. An attribute called \code{"ChartData"} also
    contains the labeled design.
}
\description{
Creates a \code{htmlwidget} summary of diagnostic information for a
text analysis output from flipTextAnalysis.
}
\seealso{
\code{\link[rhtmlMetro]{Box}}
}
