% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiontable.R
\name{CrosstabInteractionTable}
\alias{CrosstabInteractionTable}
\title{CrosstabInteractionTable}
\usage{
CrosstabInteractionTable(
  coef,
  coef.tstat,
  coef.pval,
  group.size,
  footer = "",
  title = "",
  subtitle = "",
  decimals = 2,
  p.cutoff = 0.05
)
}
\arguments{
\item{coef}{Matrix containing estimates of the coefficient.}

\item{coef.tstat}{Matrix containing t-statistics of each coefficient. Used to determine color of cell.}

\item{coef.pval}{Matrix containing p-value of each coefficient. Used to determine bolding.}

\item{group.size}{Vector containing the size of each group (i.e. interaction term).}

\item{footer}{Text to place in the footer of the table.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{decimals}{Number of decimals reported for the regression coefficients.}

\item{p.cutoff}{The alpha level for determining the significance of the coefficients.}
}
\description{
Creates pretty table showing regression coefficients with a crosstab interaction.
}
