% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetmergingbycasewidget.R
\name{DataSetMergingByCaseWidget}
\alias{DataSetMergingByCaseWidget}
\title{Widget Output for Data Set Merging By Case}
\usage{
DataSetMergingByCaseWidget(
  input.data.sets.metadata,
  merged.data.set.metadata,
  matched.names,
  merged.names,
  omitted.variable.names.list,
  input.value.attributes.list,
  is.saved.to.cloud,
  page = NULL,
  variables.per.page = 1000
)
}
\arguments{
\item{input.data.sets.metadata}{An object containing metadata for the
input data set. See the function \code{metadataFromDataSets} in flipData
for more information.}

\item{merged.data.set.metadata}{An object containing metadata for the
merged data set. See the function \code{metadataFromDataSet} in flipData
for more information.}

\item{matched.names}{A character matrix whose rows correspond to the
variables in the merged data set. The elements in each row correspond to
the input data sets and contain the names of the variables from the input
data sets that have been combined together to create a merged variable.
This matrix also has the attributes "is.fuzzy.match" and "matched.by".
is.fuzzy.match is a logical matrix of the same size as matched.names
indicating if an input variable was matched using fuzzy matching.
matched.by is a character matrix of the same size as matched.names
containing the strings "Variable name", "Variable label", "Value label"
and "Manual" indicating what data was used to match an input variable or
if the variable was matched manually.}

\item{merged.names}{A character vector of the names of the merged variables.
Also contains the attribute "renamed.variables" which is a character
matrix whose rows correspond to merged variables that have been renamed.
Each row has two elements containing the original and new name for the
variable.}

\item{omitted.variable.names.list}{A list whose elements correspond to the
input data sets. Each element contains the names of variables from a data
set that were omitted from the merged data set.}

\item{input.value.attributes.list}{A list whose elements correspond to the
variables in the merged data set. Each element is another list whose
elements correspond to the input data sets, which each of these elements
containing a named numeric vector representing the values and value labels
of a categorical input variable. This is NULL if the input variable is
not categorical.}

\item{is.saved.to.cloud}{A logical scalar indicating whether the merged data
set was saved to the Displayr cloud drive.}

\item{page}{A numeric scalar of the page number. If not specified, the first
page is shown.}

\item{variables.per.page}{A numeric scalar of the number of variables per
page.}
}
\description{
Widget shown in output for \code{flipData::MergeDataSetsByCase}.
}
