% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetmergingbyvariablewidget.R
\name{DataSetMergingByVariableWidget}
\alias{DataSetMergingByVariableWidget}
\title{Widget Output for Data Set Merging By Variable}
\usage{
DataSetMergingByVariableWidget(
  input.data.sets.metadata,
  merged.data.set.metadata,
  omitted.variables,
  is.saved.to.cloud
)
}
\arguments{
\item{input.data.sets.metadata}{An object containing metadata for the
input data set.}

\item{merged.data.set.metadata}{An object containing metadata for the
merged data set.}

\item{omitted.variables}{A list where each element contains the names of the
variables omitted from a data set.}

\item{is.saved.to.cloud}{Whether the merged data set was saved to the
Displayr cloud drive.}
}
\description{
Widget shown in output for \code{flipData::MergeDataSetsByVariable}.
}
