% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityextractionwidget.R
\name{EntityExtractionWidget}
\alias{EntityExtractionWidget}
\title{Display entity extraction output as an \code{htmlwidget}}
\usage{
EntityExtractionWidget(entity.percentages, variant.percentages, title,
  footer)
}
\arguments{
\item{entity.percentages}{The percentage of cases for each entity.}

\item{variant.percentages}{The percentage for each variant.}

\item{title}{The title to show at the top.}

\item{footer}{Footer to show containing sample information.}
}
\value{
An \code{htmlwidget} containing tables showing the output from an
  entity extraction.
}
\description{
Creates a \code{htmlwidget} summary of information for an
entity extraction output from flipTextAnalysis.
}
\seealso{
\code{\link[rhtmlMetro]{Box}}
}
