% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractCommonPrefix}
\alias{extractCommonPrefix}
\title{\code{extractCommonPrefix}}
\usage{
extractCommonPrefix(labels)
}
\arguments{
\item{labels}{A vector of labels from which we plan to extract a common prefix.}
}
\description{
\code{extractCommonPrefix}
}

