% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ExtractCommonPrefix}
\alias{ExtractCommonPrefix}
\title{\code{RemoveParentName}}
\usage{
ExtractCommonPrefix(labels)
}
\arguments{
\item{labels}{A vector of labels from which we plan to extract a common prefix.}
}
\value{
A list with components
\itemize{
\item \code{common.prefix} - The common prefix shared among the labels,
or \code{NA} if none exists
\item \code{shortened.labels} - vector of labels with the common prefix removed.
}
}
\description{
\code{ExtractCommonPrefix}
}
\details{
Returns the common prefix of a vector of labels.
}
