% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histtable.R
\name{HistTable}
\alias{HistTable}
\title{HistTable}
\usage{
HistTable(data.values, title = "", subtitle = "", footer = "", ...)
}
\arguments{
\item{data.values}{A dataframe, with each column containing the values to create a histogram.}

\item{title}{The title of the table.}

\item{subtitle}{The subtitle of the table.}

\item{footer}{The footer of the table.}

\item{...}{Additional columns to add to the table.}
}
\description{
Creates a pretty formattable table showing histograms
}
\examples{
dat <- data.frame(A=rpois(500,5), B=rpois(500,50), C=rpois(500,20))
print(HistTable(dat, 'Mean Probability'=c(5,50,100)))
}
