% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancetable.R
\name{ImportanceTable}
\alias{ImportanceTable}
\title{Create formattable table of importance scores}
\usage{
ImportanceTable(
  driver.analysis.output,
  row.labels,
  title = "",
  subtitle = "",
  footer = "",
  p.cutoff = 0.05,
  output.type
)
}
\arguments{
\item{driver.analysis.output}{A list containing the relative importance scores, output coefficients
  and significance test values. The list needs to contain the following elements:
  \itemize{
  \item \code{importance}: A numeric vector of relative importance values based off the \code{raw.importance.scores}
  \item \code{raw.importance.scores}: A numeric vector of coefficients, these could be the output coefficients from a Shapley
   Regression or Relative Importance Analysis or the Jaccard Coefficients or Correlation coefficients if the \code{output.type} (see below) is
   \code{"Shapley Regression"}, \code{"Relative Importance Analysis"}, \code{"Jaccard Coefficient"} or \code{"Correlation"}
   respectively.
    \item \code{p.values}: The resulting p values from the statistics (see below).
   }

   For \code{"Shapley Regression"} and \code{"Relative Importance Analysis"}, two more list elements are required,
   \itemize{
   \item \code{standard.errors}: The standard errors of the estimated coefficients
   \item \code{statistics}: The computed statistics
   }
   For \code{"Jaccard Coefficient"} or \code{"Correlation"} outputs, the following single extra elements is required instead.
   \itemize{
    \item \code{sample.size} The sample size count used in the calculation of significance tests for each driver.
  }}

\item{row.labels}{The row labels corresponding to the predictor variable names/labels.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}

\item{output.type}{A character string denoting the output type
\itemize{
  \item \code{"Relative Importance Analysis"}
  \item \code{"Shapley Regression"}
  \item \code{"Jaccard Coefficient"}
  \item \code{"Correlation"}
}}
}
\description{
Creates a pretty formattable table for relative/Shapley importance output.
}
\references{
This is based on code written by Kenton Russell.
}
