% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancetable.R
\name{ImportanceTable}
\alias{ImportanceTable}
\title{Create formattable table of importance scores}
\usage{
ImportanceTable(
  driver.analysis.output,
  row.labels,
  title = "",
  subtitle = "",
  footer = "",
  p.cutoff = 0.05,
  output.type
)
}
\arguments{
\item{driver.analysis.output}{A list containing the
  relative importance scores and bootstrap computed p-values. relevant elements to be plotted.
This covers the case of outputs of type
\itemize{
  \item \code{"Relative Importance Analysis"} and \code{"Shapley Regression"} where a list of importance
scores is required.
  \item \code{"Jaccard Coefficient"} where a list is required that includes the Jaccard Coefficients,
  \item \code{"Correlation"} where a list of correlation and relative importance scores and computed p values.
}}

\item{row.labels}{The row labels corresponding to the predictor variable names/labels.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}
}
\description{
Creates a pretty formattable table for relative/Shapley importance output.
}
\references{
This is based on code written by Kenton Russell.
}
