% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{Labels}
\alias{Labels}
\title{\code{Labels}}
\usage{
Labels(x, names.to.lookup = NULL, show.name = FALSE)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{names.to.lookup}{An optional list of The names of the variables or coefficients (e.g., Q2Cola for a factor).}

\item{show.name}{If \code{TRUE}, the name will prefix the extended label (where they are distinct). Ignored if
\code{names.to.lookup} is provided.}
}
\value{
A \code{vector} of labels
}
\description{
Replaces a list of given names or coefficient, with any underlying labels. If
}
\details{
First tries to find the "label" attribute, then "name", then "question", and lastly looks to the variable's name.
Where \code{names.to.lookup} is provided, Works for dummy variables as well as normal variables.
Trims backticks and whitespace. Returns names where labels cannot be found.
}

