% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{Labels}
\alias{Labels}
\title{\code{Labels}}
\usage{
Labels(x, names = NULL)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{names}{An optional list of The names of the variables or coefficients (e.g., Q2Cola for a factor).}
}
\value{
A \code{vector} of labels
}
\description{
Replaces a list of given names or coefficient, with any underlying labels.
}
\details{
Returns the names if the labels to not exist.
Where \code{names} is provided, Works for dummy variables as well as normal variables.
Trims backticks and whitespace. Returns names where labels cannot be found.
}

