% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeimportancetable.R
\name{RelativeImportanceTable}
\alias{RelativeImportanceTable}
\title{RelativeImportanceTable}
\usage{
RelativeImportanceTable(relative.importance, row.labels, title = "",
  subtitle = "", footer = "", p.cutoff = 0.05)
}
\arguments{
\item{relative.importance}{The relative importance object.}

\item{row.labels}{The row labels corresponding to the predictor variable names/labels.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}

\item{footer}{Text to place in the footer of the table.}

\item{p.cutoff}{The alpha level used when formatting the p-value column.}
}
\description{
Creates a pretty formattable table for relative importance output.
}
\references{
This is based on code written by Kenton Russell.
}
