% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significancetest.R
\name{SignificanceTest}
\alias{SignificanceTest}
\title{\code{SignificanceTest}}
\usage{
SignificanceTest(obj, test.name, var1, var2 = NULL, filter, weight,
  p.value.method = "", show.labels = TRUE, decimal.places = NULL)
}
\arguments{
\item{obj}{Significance testing output object, e.g. object of class htest.}

\item{test.name}{Name of the test.}

\item{var1}{Variable 1 used in test.}

\item{var2}{Variable 2 used in test.}

\item{filter}{Filter variable.}

\item{weight}{Weight variable.}

\item{p.value.method}{Specifies how the p-value is computed.}

\item{show.labels}{Whether to show variable labels instead of names.}

\item{decimal.places}{The number of decimal places to show.}
}
\description{
Standardize information to be shown in significance test results.
}

