% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significancetest.R
\name{SignificanceTest}
\alias{SignificanceTest}
\title{SignificanceTest}
\usage{
SignificanceTest(obj, test.name, vars = NULL, filter = NULL,
  weight = NULL, p.value.method = "", show.labels = TRUE,
  decimal.places = NULL, missing = "Exclude cases with missing data",
  reg.name = NULL, reg.sample.description = NULL, resample = FALSE,
  group.levels = NULL)
}
\arguments{
\item{obj}{Significance testing output object, e.g. object of class htest.}

\item{test.name}{Name of the test.}

\item{vars}{Variables used in test.}

\item{filter}{Filter variable.}

\item{weight}{Weight variable.}

\item{p.value.method}{Specifies how the p-value is computed.}

\item{show.labels}{Whether to show variable labels instead of names.}

\item{decimal.places}{The number of decimal places to show.}

\item{missing}{How missing data is to be treated in the analysis.
Options are: \code{"Error if missing data"}, \code{"Exclude cases with missing data"} and
\code{"Imputation (replace missing values with estimates)"}.}

\item{reg.name}{The name of the regression object on which the test is run.}

\item{reg.sample.description}{The sample description of the regression object on which the test is run.}

\item{resample}{Whether resampling is used whenever weights are applied.}

\item{group.levels}{The levels of the categorical group variable.}
}
\description{
Standardize information to be shown in significance test results.
}
\details{
This function was created for the following Standard R pages:
\itemize{
\item{Missing Data - Little's MCAR Test}
\item{Regression - Diagnostic - Heteroscedasticity}
\item{Regression - Diagnostic - Normality (Shapiro-Wilk)}
\item{Regression - Diagnostic - Serial Correlation (Durbin-Watson)}
\item{Test - Bartlett Test of Sphericity}
\item{Test - Chi-Square Test of Independence}
\item{Test - Correlation}
\item{Test - Nonparametric - Kruskal-Wallis Rank Sum Test}
\item{Test - Nonparametric - Paired Samples Wilcoxon Test}
\item{Test - Nonparametric - Single-Sample Wilcoxon Test}
\item{Test - Nonparametric - Two-Sample Wilcoxon Rank Sum Test}
\item{Test - Variance - F-Test to Compare Two Variances}
}
}
