% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingwidget.R
\name{StackingWidget}
\alias{StackingWidget}
\title{Stacking Output Widget}
\usage{
StackingWidget(
  stacked.data.set.metadata,
  unstackable.names,
  omitted.variables,
  omitted.stacked.variables,
  common.labels,
  is.saved.to.cloud
)
}
\arguments{
\item{stacked.data.set.metadata}{An object containing metadata for the
stacked data set.}

\item{unstackable.names}{A list of character vectors of the names of
variables that could not be stacked due to mismatching types or
value attributes.}

\item{omitted.variables}{A character vector of omitted variables.}

\item{omitted.stacked.variables}{A character vector of omitted stacked
variables.}

\item{common.labels}{The common labels used for stacking.}

\item{is.saved.to.cloud}{Whether the stacked data set was saved to the
Displayr cloud drive.}
}
\description{
Stacking widget shown in output for \code{flipData::StackData}.
}
