% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textclassifierwidget.R
\name{TextClassifierWidget}
\alias{TextClassifierWidget}
\title{Display text classifier output as an \code{htmlwidget}}
\usage{
TextClassifierWidget(observed.counts, predicted.counts, category.accuracy,
  examples, overall.metrics, cv.metrics, text.raw.by.categorization,
  missing, title, footer)
}
\arguments{
\item{observed.counts}{Numeric vector of counts of the observed existing categories}

\item{predicted.counts}{Numeric vector of counts of the predicted categories}

\item{category.accuracy}{Numeric vector of accuracy over training sample.}

\item{examples}{Examples for each category.}

\item{category.accuracy}{Numeric vector of accuracy over training sample.}

\item{overall.metrics}{Numeric vector of the overall performance metrics on the trained model.}

\item{cv.metrics}{Numeric matrix of the performance metrics on the cross validation data.}

\item{text.raw.by.categorization}{A list containing the raw text for each
category.}

\item{missing}{Logical vector indicating which cases are missing.}

\item{title}{The title to show at the top.}

\item{footer}{Footer to show containing sample information.}
}
\value{
An \code{htmlwidget} containing tables showing the output from an
  automatic text categorization.
}
\description{
Creates a \code{htmlwidget} summary of information for a
text classifier output from flipTextAnalysis.
}
\seealso{
\code{\link[rhtmlMetro]{Box}}
}
